Added show endpoint for User (#6)
* added new show endpoint

* fix server crash when bad requests are sent

* removed role from returned attributes

* safe ternary

+++ src/controller/UserController.js
@@ -1,17 +1,37 @@
-const jwt = require("jsonwebtoken");
 const bcrypt = require("bcrypt");
 const UserModel = require("./../models/User");
-const Role = require("../enum/RolesEnum");
+const { getAuthenticatedUser, isUserOwner, isUserAdmin } = require("../utils/Security/SecurityHelper");
 
 module.exports = {
+    show: async (req, res) => {
+        try {
+            const user = req.params.id ? await UserModel.findById(req.params.id) : null;
+            const authenticatedUser = await getAuthenticatedUser(req);
+
+            if (!isUserOwner(authenticatedUser, user)) {
+                return res.status(403).send({
+                    message: "Vous n'êtes pas autorisé à voir les informations de cet utilisateur.",
+                });
+            }
+
+            res.send({
+                firstname: user.firstname,
+                lastname: user.lastname,
+                email: user.email,
+            });
+        } catch (error) {
+            res.status(500).send({
+                message: error.message || "Erreur 500 : Impossible de récupérer les informations de l'utilisateur",
+            });
+        }
+    },
+
     edit: async (req, res) => {
         try {
-            let token = req.headers["authorization"].replace("Bearer ", "");
-            const { userId } = jwt.verify(token, process.env.JWT_SECRET || "secret");
-            const requestingUser = await UserModel.findById(userId);
-            const user = await UserModel.findById(req.params.id);
+            const user = req.params.id ? await UserModel.findById(req.params.id) : null;
+            const authenticatedUser = await getAuthenticatedUser(req);
 
-            if (requestingUser.role !== Role.ADMIN && requestingUser._id.toString() !== user._id.toString()) {
+            if (!isUserOwner(authenticatedUser, user)) {
                 return res.status(403).send({
                     message: "Vous n'êtes pas autorisé à changer les informations de cet utilisateur.",
                 });
@@ -20,9 +40,11 @@ module.exports = {
             user.firstname = req.body.firstname || user.firstname;
             user.lastname = req.body.lastname || user.lastname;
             user.email = req.body.email || user.email;
+
             if (req.body.password) {
                 user.password = await bcrypt.hash(req.body.password, 10);
             }
+
             user.role = req.body.role || user.role;
 
             await user.save();
@@ -45,6 +67,14 @@ module.exports = {
     },
 
     delete: async (req, res) => {
+        const authenticatedUser = await getAuthenticatedUser(req);
+
+        if (!isUserAdmin(authenticatedUser)) {
+            return res.status(403).send({
+                message: "Vous n'êtes pas autorisé à supprimer un utilisateur",
+            });
+        }
+
         try {
             const user = await UserModel.findByIdAndDelete(req.params.id);
             if (!user) {
